Docs, improve TaskProgress handling
