+++ engine/launcher/src/main/scala/org/enso/launcher/releases/launcher/LauncherRelease.scala
@@ -3,7 +3,7 @@ package org.enso.launcher.releases.launcher
 import java.nio.file.Path
 
 import nl.gn0s1s.bump.SemVer
-import org.enso.cli.TaskProgress
+import org.enso.cli.task.TaskProgress
 import org.enso.runtimeversionmanager.CurrentVersion
 
 /** Represents a launcher release.
