+++ engine/launcher/src/test/scala/org/enso/launcher/releases/fallback/TestFileStorage.scala
@@ -2,7 +2,7 @@ package org.enso.launcher.releases.fallback
 
 import java.nio.file.Path
 
-import org.enso.cli.{TaskProgress, TaskProgressImplementation}
+import org.enso.cli.task.{TaskProgress, TaskProgressImplementation}
 import org.enso.runtimeversionmanager.FileSystem
 import org.enso.launcher.TestHelpers
 import org.enso.launcher.releases.fallback.staticwebsite.FileStorage
