+++ lib/scala/cli/src/main/scala/org/enso/cli/task/TaskProgress.scala
@@ -1,15 +1,8 @@
-package org.enso.cli
+package org.enso.cli.task
 
 import java.util.concurrent.LinkedTransferQueue
 
 import scala.util.{Failure, Try}
-
-/** Clients can implement this trait to get progress updates.
-  */
-trait ProgressListener[-A] {
-  def progressUpdate(done: Long, total: Option[Long]): Unit
-  def done(result: Try[A]):                            Unit
-}
 
 /** Represents a long-running background task.
   */
@@ -23,17 +16,6 @@ trait TaskProgress[A] {
     * often [[ProgressListener.progressUpdate]] is called.
     */
   def addProgressListener(listener: ProgressListener[A]): Unit
-
-  /** Blocks and waits for the completion of the task.
-    *
-    * Optionally displays a progress bar in the terminal. Returns a [[Try]]
-    * value that wraps the result.
-    *
-    * @param showProgress whether or not to show a progress bar while waiting
-    */
-  def waitForResult(showProgress: Boolean = false): Try[A] =
-    if (showProgress) ProgressBar.waitWithProgress(this)
-    else TaskProgress.waitForTask(this)
 
   /** Waits for the task to finish and returns its value, throwing any
     * exceptions that were reported.
@@ -100,74 +82,5 @@ object TaskProgress {
     })
 
     queue.take()
-  }
-}
-
-/** Transforms the result of the `source` task with `f`.
-  *
-  * Used internally by [[TaskProgress.flatMap]].
-  */
-private class MappedTask[A, B](source: TaskProgress[A], f: A => Try[B])
-    extends TaskProgress[B] {
-  override def addProgressListener(
-    listener: ProgressListener[B]
-  ): Unit =
-    source.addProgressListener(new ProgressListener[A] {
-      override def progressUpdate(done: Long, total: Option[Long]): Unit =
-        listener.progressUpdate(done, total)
-      override def done(result: Try[A]): Unit =
-        listener.done(result.flatMap(f))
-    })
-}
-
-/** A simple implementation of [[TaskProgress]] that can be used to report
-  * progress updates and mark completion of a task.
-  */
-class TaskProgressImplementation[A] extends TaskProgress[A] {
-  @volatile private var listeners: List[ProgressListener[A]] = Nil
-  private var result: Option[Try[A]]                         = None
-
-  override def addProgressListener(
-    listener: ProgressListener[A]
-  ): Unit = {
-    this.synchronized {
-      result match {
-        case Some(value) =>
-          listener.done(value)
-        case None =>
-      }
-
-      listeners ::= listener
-    }
-  }
-
-  /** Marks the completion of this task.
-    *
-    * All registered listeners are immediately notified. Any listeners added
-    * later will also be notified as soon as they are added.
-    *
-    * Can be called only once per task.
-    * @param result the result to complete the task with
-    */
-  def setComplete(result: Try[A]): Unit = {
-    this.synchronized {
-      if (this.result.isDefined) {
-        throw new IllegalStateException(
-          "A task has been completed more than once."
-        )
-      }
-
-      this.result = Some(result)
-      listeners.foreach(_.done(result))
-    }
-  }
-
-  /** Report a progress update to all registered listeners.
-    *
-    * This operation is not synchronized, as it is not a problem if a just added
-    * listener does not get the latest progress update.
-    */
-  def reportProgress(done: Long, total: Option[Long]): Unit = {
-    listeners.foreach(_.progressUpdate(done, total))
   }
 }
