+++ lib/scala/project-manager/src/main/scala/org/enso/projectmanager/service/versionmanagement/NoOpInterface.scala
@@ -1,7 +1,7 @@
 package org.enso.projectmanager.service.versionmanagement
 
 import nl.gn0s1s.bump.SemVer
-import org.enso.cli.TaskProgress
+import org.enso.cli.task.TaskProgress
 import org.enso.runtimeversionmanager.components.{
   GraalVMVersion,
   RuntimeVersionManagementUserInterface
