+++ lib/scala/runtime-version-manager/src/main/scala/org/enso/runtimeversionmanager/releases/github/GithubAsset.scala
@@ -2,7 +2,7 @@ package org.enso.runtimeversionmanager.releases.github
 
 import java.nio.file.Path
 
-import org.enso.cli.TaskProgress
+import org.enso.cli.task.TaskProgress
 import org.enso.runtimeversionmanager.releases.Asset
 
 case class GithubAsset(asset: GithubAPI.Asset) extends Asset {
