+++ lib/scala/runtime-version-manager/src/main/scala/org/enso/runtimeversionmanager/releases/graalvm/GraalVMRuntimeReleaseProvider.scala
@@ -2,7 +2,7 @@ package org.enso.runtimeversionmanager.releases.graalvm
 
 import java.nio.file.Path
 
-import org.enso.cli.TaskProgress
+import org.enso.cli.task.TaskProgress
 import org.enso.runtimeversionmanager.components.GraalVMVersion
 
 /** Interface for a service providing GraalVM runtime releases. */
