+++ lib/scala/runtime-version-manager/src/main/scala/org/enso/runtimeversionmanager/releases/testing/FakeReleaseProvider.scala
@@ -2,7 +2,7 @@ package org.enso.runtimeversionmanager.releases.testing
 
 import java.nio.file.{Files, Path, StandardCopyOption}
 
-import org.enso.cli.{ProgressListener, TaskProgress}
+import org.enso.cli.task.{ProgressListener, TaskProgress}
 import org.enso.runtimeversionmanager.locking.{LockManager, LockType}
 import org.enso.runtimeversionmanager.releases.{
   Asset,
